+++ packages/app/lib/models/browser/browser_favorite.g.dart
@@ -88,7 +88,7 @@ const BrowserFavoriteSchema = CollectionSchema(
   getId: _browserFavoriteGetId,
   getLinks: _browserFavoriteGetLinks,
   attach: _browserFavoriteAttach,
-  version: '3.1.0+1',
+  version: '3.1.8',
 );
 
 int _browserFavoriteEstimateSize(