+++ packages/app/lib/nostr-core/relay_websocket.dart
@@ -8,9 +8,9 @@ import 'package:app/service/websocket.service.dart';
 import 'package:app/utils.dart';
 import 'package:async_queue/async_queue.dart';
 import 'package:easy_debounce/easy_throttle.dart';
+import 'package:keychat_ecash/NostrWalletConnect/NostrWalletConnect_controller.dart';
 import 'package:web_socket_channel/io.dart';
 import 'package:web_socket_channel/web_socket_channel.dart';
-import '../utils.dart' as utils;
 
 const _maxReqCount = 20; // max pool size is 32. be setting by relay server
 
@@ -43,10 +43,10 @@ class RelayWebsocket {
 
   Future listenPubkeys(List<String> pubkeys, DateTime since,
       [List<int> kinds = const [EventKinds.encryptedDirectMessage]]) async {
-    List<List<String>> groups = utils.listToGroupList(pubkeys, 120);
+    List<List<String>> groups = listToGroupList(pubkeys, 120);
 
     for (List<String> group in groups) {
-      String subId = utils.generate64RandomHexChars(16);
+      String subId = generate64RandomHexChars(16);
 
       NostrNip4Req req = NostrNip4Req(
           reqId: subId, kinds: kinds, pubkeys: group, since: since);
@@ -79,7 +79,7 @@ class RelayWebsocket {
 
   _statusCheck() {
     if (channel == null || channelStatus != RelayStatusEnum.success) {
-      throw Exception('disconnected_${relay.url}');
+      throw Exception('disconnected: ${relay.url}');
     }
   }
 
@@ -144,6 +144,9 @@ class RelayWebsocket {
     _startListen();
     await Future.delayed(const Duration(seconds: 1));
     _proccessFailedEvents();
+    // nwc reconnect
+    Utils.getGetxController<NostrWalletConnectController>()
+        ?.startListening(relay.url);
   }
 
   void connecting() {